pyproject.toml
[project]
name = "blogging"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "flask>=3.1.2",
    "flask-bcrypt>=1.0.1",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "markdown>=3.10",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.6",
    "wtforms>=3.2.1",
]
